Apply review suggestions
